requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
fastapi==0.115.6
Flask==3.1.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==308
pyzmq==26.2.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
spotipy==2.24.0
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
